requirements_gemini_live.txt
google-genai>=1.0.0
livekit-agents>=1.5.0
livekit>=0.19.0
httpx>=0.27.0
numpy>=1.24.0
